+++ Applications/Monet/MAIntonationView.m
@@ -176,9 +176,9 @@ NSString *MAIntonationViewSelectionDidChangeNotification = @"MAIntonationViewSel
     nonretained_delegate = newDelegate;
 }
 
-- (float)minimumWidth;
+- (CGFloat)minimumWidth;
 {
-    float minimumWidth;
+    CGFloat minimumWidth;
 
     if ([[eventList events] count] == 0) {
         minimumWidth = 0.0;
@@ -257,8 +257,8 @@ NSString *MAIntonationViewSelectionDidChangeNotification = @"MAIntonationViewSel
     NSBezierPath *bezierPath;
     NSRect bounds;
     NSPoint graphOrigin;
-    int sectionHeight;
-    int index;
+    CGFloat sectionHeight;
+    NSUInteger index;
 
     bounds = NSIntegralRect([self bounds]);
     graphOrigin = [self graphOrigin];
@@ -317,7 +317,7 @@ NSString *MAIntonationViewSelectionDidChangeNotification = @"MAIntonationViewSel
     NSRect bounds, rect;
     NSBezierPath *bezierPath;
     NSPoint point;
-    float time;
+    CGFloat time;
     NSRect cellFrame;
 
     bounds = NSIntegralRect([self bounds]);
@@ -343,7 +343,7 @@ NSString *MAIntonationViewSelectionDidChangeNotification = @"MAIntonationViewSel
         point.y = graphOrigin.y;
         [bezierPath moveToPoint:point];
 
-        if (((int)time % 100) == 0) {
+        if (((NSInteger)time % 100) == 0) {
             point.y -= 10;
 
             [labelTextFieldCell setIntValue:time];
@@ -376,11 +376,11 @@ NSString *MAIntonationViewSelectionDidChangeNotification = @"MAIntonationViewSel
 // TODO (2004-08-16): Need right margin so that the last posture is visible.
 - (void)drawPostureLabels;
 {
-    int count, index;
+    NSUInteger count, index;
     MMPosture *currentPosture;
     NSRect bounds;
-    float currentX;
-    int postureIndex = 0;
+    CGFloat currentX;
+    NSUInteger postureIndex = 0;
     NSArray *events;
 
     bounds = NSIntegralRect([self bounds]);
@@ -409,11 +409,11 @@ NSString *MAIntonationViewSelectionDidChangeNotification = @"MAIntonationViewSel
 - (void)drawRules;
 {
     NSBezierPath *bezierPath;
-    float currentX, extraWidth;
-    int count, index;
+    CGFloat currentX, extraWidth;
+    NSUInteger count, index;
     NSRect bounds;
     NSPoint graphOrigin;
-    int sectionHeight;
+    CGFloat sectionHeight;
     struct _rule *rule;
 
     bounds = NSIntegralRect([self bounds]);
@@ -471,11 +471,11 @@ NSString *MAIntonationViewSelectionDidChangeNotification = @"MAIntonationViewSel
 
 - (void)drawRuleBackground;
 {
-    float extraWidth;
-    int count, index;
+    CGFloat extraWidth;
+    NSUInteger count, index;
     NSRect ruleFrame;
     NSPoint graphOrigin;
-    int sectionHeight;
+    CGFloat sectionHeight;
     struct _rule *rule;
 
     graphOrigin = [self graphOrigin];
@@ -505,7 +505,7 @@ NSString *MAIntonationViewSelectionDidChangeNotification = @"MAIntonationViewSel
 - (void)drawIntonationPoints;
 {
     NSBezierPath *bezierPath;
-    int count, index;
+    NSUInteger count, index;
     NSPoint currentPoint;
     NSRect bounds;
     NSPoint graphOrigin;
@@ -552,7 +552,7 @@ NSString *MAIntonationViewSelectionDidChangeNotification = @"MAIntonationViewSel
     double x2, y2, m2, x22, x23;
     double denominator;
     double x, y;
-    int i, j;
+    NSUInteger i, j;
     id point1, point2;
     NSBezierPath *bezierPath;
     NSArray *intonationPoints = [eventList intonationPoints];
@@ -624,7 +624,7 @@ NSString *MAIntonationViewSelectionDidChangeNotification = @"MAIntonationViewSel
 #ifdef PORTING
     NSEvent *nextEvent;
     NSPoint position;
-    int time;
+    CGFloat time;
 
     [[self window] setAcceptsMouseMovedEvents:YES];
     while (1) {
@@ -654,17 +654,17 @@ NSString *MAIntonationViewSelectionDidChangeNotification = @"MAIntonationViewSel
 - (void)keyDown:(NSEvent *)keyEvent;
 {
     NSString *characters;
-    int index, length;
+    NSUInteger index, length;
     unichar ch;
 
-    unsigned int ruleCount;
+    NSUInteger ruleCount;
 
     ruleCount = [eventList ruleCount];
 
     characters = [keyEvent characters];
     length = [characters length];
     for (index = 0; index < length; index++) {
-        unsigned int pointCount, pointIndex;
+        NSUInteger pointCount, pointIndex;
 
         pointCount = [selectedPoints count];
 
@@ -759,8 +759,8 @@ NSString *MAIntonationViewSelectionDidChangeNotification = @"MAIntonationViewSel
         if ([mouseEvent modifierFlags] & NSAlternateKeyMask) {
             MMIntonationPoint *newIntonationPoint;
             NSPoint graphOrigin;
-            int yScale, absoluteTime;
-            int ruleIndex;
+            CGFloat yScale, absoluteTime;
+            NSUInteger ruleIndex;
             double offsetTime;
 
             graphOrigin = [self graphOrigin];
@@ -819,9 +819,9 @@ NSString *MAIntonationViewSelectionDidChangeNotification = @"MAIntonationViewSel
 {
     NSPoint graphOrigin;
     NSRect selectionRect;
-    int count, index;
+    NSUInteger count, index;
     //float timeScale;
-    int yScale;
+    CGFloat yScale;
     NSArray *intonationPoints;
     NSRect bounds;
 
@@ -890,14 +890,14 @@ NSString *MAIntonationViewSelectionDidChangeNotification = @"MAIntonationViewSel
 // Double-clicking adds intonation point?
 - (void)mouseDown:(NSEvent *)theEvent;
 {
-    float row, column;
-    float row1, column1;
-    float row2, column2;
-    float temp, distance, distance1, tally = 0.0, tally1 = 0.0;
-    float semitone;
+    CGFloat row, column;
+    CGFloat row1, column1;
+    CGFloat row2, column2;
+    CGFloat temp, distance, distance1, tally = 0.0, tally1 = 0.0;
+    CGFloat semitone;
     NSPoint mouseDownLocation = [theEvent locationInWindow];
     NSEvent *newEvent;
-    int i, ruleIndex = 0;
+    NSUInteger i, ruleIndex = 0;
     struct _rule *rule;
     MMIntonationPoint *iPoint;
     id tempPoint;
@@ -942,7 +942,7 @@ NSString *MAIntonationViewSelectionDidChangeNotification = @"MAIntonationViewSel
 #ifdef PORTING
     NSPoint mouseDownLocation;
     NSEvent *newEvent;
-    float delta, originalScale;
+    CGFloat delta, originalScale;
 
     originalScale = timeScale;
 
@@ -1014,10 +1014,10 @@ NSString *MAIntonationViewSelectionDidChangeNotification = @"MAIntonationViewSel
 // View geometry
 //
 
-- (int)sectionHeight;
+- (CGFloat)sectionHeight;
 {
     NSRect bounds;
-    int sectionHeight;
+    CGFloat sectionHeight;
 
     bounds = [self bounds];
     sectionHeight = (bounds.size.height - TOP_MARGIN - BOTTOM_MARGIN) / SECTION_COUNT;
@@ -1041,19 +1041,19 @@ NSString *MAIntonationViewSelectionDidChangeNotification = @"MAIntonationViewSel
     [self setNeedsDisplay:YES];
 }
 
-- (float)scaleXPosition:(float)xPosition;
+- (CGFloat)scaleXPosition:(CGFloat)xPosition;
 {
     return floor(xPosition / timeScale);
 }
 
-- (float)scaleWidth:(float)width;
+- (CGFloat)scaleWidth:(CGFloat)width;
 {
     return floor(width / timeScale);
 }
 
 - (NSRect)rectFormedByPoint:(NSPoint)point1 andPoint:(NSPoint)point2;
 {
-    float minx, miny, maxx, maxy;
+    CGFloat minx, miny, maxx, maxy;
     NSRect rect;
 
     if (point1.x < point2.x) {
@@ -1080,7 +1080,7 @@ NSString *MAIntonationViewSelectionDidChangeNotification = @"MAIntonationViewSel
     return rect;
 }
 
-- (float)convertYPositionToSemitone:(float)yPosition;
+- (CGFloat)convertYPositionToSemitone:(CGFloat)yPosition;
 {
     NSPoint graphOrigin;
 
@@ -1089,7 +1089,7 @@ NSString *MAIntonationViewSelectionDidChangeNotification = @"MAIntonationViewSel
     return ((yPosition - graphOrigin.y) / [self sectionHeight]) - ZERO_SECTION;
 }
 
-- (float)convertXPositionToTime:(float)xPosition;
+- (CGFloat)convertXPositionToTime:(CGFloat)xPosition;
 {
     return xPosition * timeScale;
 }
@@ -1102,7 +1102,7 @@ NSString *MAIntonationViewSelectionDidChangeNotification = @"MAIntonationViewSel
 
 - (void)removeOldSelectedPoints;
 {
-    int count, index;
+    NSUInteger count, index;
 
     // This is another case where count cannot be unsigned.
     count = [selectedPoints count];
