+++ Applications/Monet/MMDisplayParameter.m
@@ -42,12 +42,12 @@
     isSpecial = newFlag;
 }
 
-- (int)tag;
+- (NSUInteger)tag;
 {
     return tag;
 }
 
-- (void)setTag:(int)newTag;
+- (void)setTag:(NSUInteger)newTag;
 {
     tag = newTag;
 }
